backend/pyproject.toml
[project]
name = "masterformat-pdf-parser"
version = "1.0.0"
description = "Backend API for parsing MasterFormat PDFs to structured JSON using Gemini LLM"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.3.0",
    "PyPDF2>=3.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyMuPDF>=1.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 